requirements.txt
anthropic>=0.21.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
tabulate>=0.9.0
rich>=13.0.0
tqdm>=4.65.0